Moved some function calls (saveTextOuputs and prepFiles) from the server module to the util module. Changed the name of read_file to loadAudio.
